Finally suceeded in the connection of server and client

+++ OpenUsuiSuji/Form1.cs
@@ -13,11 +13,16 @@ using System.Windows.Forms;
 
 namespace OpenUsuiSuji
 {
+    delegate void ClientCallback(long t1, long t2);
+
     public partial class Form1 : Form
     {
         public Form1()
         {
             InitializeComponent();
+
+            T1Value.Paint += new PaintEventHandler(UpdateT1);
+            T2Value.Paint += new PaintEventHandler(UpdateT2);
 
             if (System.Environment.Is64BitOperatingSystem)
             {
@@ -28,23 +33,31 @@ namespace OpenUsuiSuji
                 this.Text = "x86";
             }
 
-            watchers = watcherNames.ToDictionary(name => name, name => new InputWatcher(name, pipeName));
+            watchers = watcherNames.ToDictionary(
+                key => key, 
+                key => new InputWatcher(key, GetWatcherPipeName(name, version, key)));
+
             foreach (var name in watchers.Keys)
             {
                 watchers[name].Start();
-            }   
+            }
+            
+            clients = watcherNames.ToDictionary(
+                key => key,
+                key => new Client(GetWatcherPipeName(name, version, key), UpdateDisplay));
+
+            foreach (var name in clients.Keys)
+            {
+                clients[name].Start();
+            }
         }
 
         private const String name = "OpenUsuiSuji";
         private const String version = "0.00";
 
-        private static readonly String pipeName = string.Format("{0}-{1}", name, version);
-        private NamedPipeServerStream pipe = null;
-
-
-        private void ClosePipe()
+        private String GetWatcherPipeName(String name, String version, String postfix)
         {
-            pipe.Close();    
+            return string.Format("{0}-{1}-{2}", name, version, postfix);
         }
 
         private static readonly List<String> watcherNames = 
@@ -54,6 +67,107 @@ namespace OpenUsuiSuji
             };
 
         private Dictionary<String, InputWatcher> watchers = null;
+        private Dictionary<String, Client> clients = null;
+
+
+        private long t1 = 0;
+        private long t2 = 0;
+
+        public void UpdateDisplay(long t1, long t2)
+        {
+            if (T1Value.InvokeRequired)
+            {
+                ClientCallback f = new ClientCallback(UpdateDisplay);
+                this.Invoke(f, new object[] { t1, t2 });
+            } 
+            else
+            {
+                this.t1 = t1;
+                this.t2 = t2;
+                T1Value.Refresh();
+                T2Value.Refresh();
+            }
+        }
+
+        public void UpdateT1(object sender, PaintEventArgs args)
+        {
+            T1Value.Text = t1.ToString();
+        }
+
+        public void UpdateT2(object sender, PaintEventArgs args)
+        {
+            T2Value.Text = t2.ToString();
+        }
+    }
+
+    class Client: IDisposable
+    {
+        private Thread thread = null;
+        private NamedPipeClientStream stream = null;
+        private String pipeName = null;
+        private ClientCallback callback = null;
+
+        private static readonly byte[] OK = { 0x4F, 0x4B };
+
+        public Client(String pipeName, ClientCallback callback)
+        {
+            this.pipeName = pipeName;
+            this.callback = callback;
+        }
+
+        public void Start()
+        {
+            thread = new Thread(Watch);
+            thread.Start();
+        }
+        public void Dispose()
+        {
+            if (thread != null)
+            {
+                thread.Abort();
+            }
+            
+            try
+            {
+                if (stream != null)
+                {
+                    stream.Close();
+                }
+            }
+            catch
+            {
+
+            }
+        }
+
+        private void Watch()
+        {
+            stream = new NamedPipeClientStream(".", pipeName, PipeDirection.InOut);
+            stream.Connect();
+            while (true)
+            {
+                byte[] buf = new byte[8];
+                long t1 = DateTime.Now.Ticks;
+                stream.Read(buf, 0, 8);
+                long t2 = DateTime.Now.Ticks;
+                long t0 = BitConverter.ToInt64(buf, 0);
+
+                Debug.Print("received: {0}", BitConverter.ToString(buf));
+                Debug.Print("t0: {0}", t0);
+                Debug.Print("t1: {0}", t1);
+                Debug.Print("t2: {0}", t2);
+
+                //callback(t1, t2);
+                callback(t1 - t0, t2 - t0);
+
+                stream.WriteByte(79);
+                stream.WriteByte(75);
+                stream.Flush();
+                stream.WaitForPipeDrain();
+
+                //Thread.Sleep(1000);
+            }
+        }
     }
 
     class InputWatcher: IDisposable
@@ -62,10 +176,10 @@ namespace OpenUsuiSuji
         private Process process = null;
         private String pipeName = null;
         private String name = null;
-
-        public InputWatcher(String watcherName, String pipeName)
+        
+        public InputWatcher(String name, String pipeName)
         {
-            this.name = watcherName;
+            this.name = name;
             this.pipeName = pipeName;
         }
 
@@ -77,10 +191,16 @@ namespace OpenUsuiSuji
 
         public void Dispose()
         {
-            thread.Abort();
+            if (thread != null)
+            {
+                thread.Abort();
+            }
             try
             {
-                process.Kill();
+                if (process != null)
+                {
+                    process.Kill();
+                }
             }
             catch
             {
@@ -90,15 +210,20 @@ namespace OpenUsuiSuji
 
         private void Watch()
         {
-            ProcessStartInfo info = new ProcessStartInfo()
+            while (true)
             {
-                FileName = name,
-                Arguments = pipeName,
-                CreateNoWindow = true,
-                UseShellExecute = false
-            };
-            process = Process.Start(info);
-            process.WaitForExit();
+                ProcessStartInfo info = new ProcessStartInfo()
+                {
+                    FileName = name,
+                    Arguments = pipeName,
+                    CreateNoWindow = true,
+                    UseShellExecute = false
+                };
+                process = Process.Start(info);
+                process.WaitForExit();
+
+                Thread.Sleep(1000);
+            }
         }
     }
 }
